Read all files through the registry
This ensures that a file is within its own registry
